requirements.txt
fastapi==0.111.0
httpx==0.27.0
jinja2==3.1.4
pydantic==2.8.2
python-docx==1.1.2
PyMuPDF==1.24.9
uvicorn==0.30.6
